Move app.dart
